Remove implicit casts

+++ src/rpi-gpio.c
@@ -8,7 +8,7 @@ void RPI_SetGpioPinFunction(rpi_gpio_pin_t gpio, rpi_gpio_alt_function_t func)
   rpi_reg_rw_t* fsel_reg = &RPI_GpioBase->GPFSEL[gpio / 10];
 
   rpi_reg_rw_t fsel_copy = *fsel_reg;
-  fsel_copy &= ~(FS_MASK << ((gpio % 10) * 3));
+  fsel_copy &= (uint32_t)~(FS_MASK << ((gpio % 10) * 3));
   fsel_copy |= (func << ((gpio % 10) * 3));
   *fsel_reg = fsel_copy;
 }
